Move mechs test to its own file
